+++ api/src/models/units/Unit.ts
@@ -11,6 +11,7 @@ import {taskUnitSchema} from './TaskUnit';
 import {IUser} from '../../../../shared/models/IUser';
 import {IProgress} from '../../../../shared/models/progress/IProgress';
 import {User} from '../User';
+import {ChatRoom} from '../ChatRoom';
 
 interface IUnitModel extends IUnit, mongoose.Document {
   exportJSON: (onlyBasicData?: boolean) => Promise<IUnit>;
@@ -49,6 +50,10 @@ const unitSchema = new mongoose.Schema({
     unitCreator: {
       type: mongoose.Schema.Types.ObjectId,
       ref: 'User'
+    },
+    chatRoom: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'ChatRoom'
     }
   },
   {
@@ -64,10 +69,28 @@ const unitSchema = new mongoose.Schema({
         if (ret._course) {
           ret._course = ret._course.toString();
         }
+
+        if (ret.hasOwnProperty('chatRoom') && ret.chatRoom) {
+          ret.chatRoom = ret.chatRoom.toString();
+        }
       }
     },
   }
 );
+
+
+unitSchema.pre('save', async function () {
+  const unit = <IUnitModel> this;
+  if (this.isNew) {
+   const chatRoom = await ChatRoom.create({
+      room: {
+        roomType: 'Unit',
+        roomFor: unit
+      }
+    });
+    this.set('chatRoom', chatRoom);
+  }
+});
 
 unitSchema.virtual('progressData', {
   ref: 'Progress',
@@ -87,7 +110,6 @@ unitSchema.methods.exportJSON = function (onlyBasicData: boolean= false) {
   delete obj.updatedAt;
   delete obj.unitCreator;
   delete obj.files;
-
 
   // custom properties
   delete obj._course;