+++ shared/models/messaging/SoketIOEvent.ts
@@ -1,0 +1,5 @@
+export enum SocketIOEvent {
+  CONNECT = 'connect',
+  DISCONNECT  = 'disconnect',
+  MESSAGE  = 'message',
+}